% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{greet}
\alias{greet}
\title{Greeting A Person}
\usage{
greet(name)
}
\arguments{
\item{name}{character() Name of person}
}
\value{
character() Greeting
}
\description{
Given a person's name, a customized greeting will be given.
}
\examples{
greet("Mary")
}
