% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFP-package.r
\docType{package}
\name{MFP}
\alias{MFP}
\alias{MFP-package}
\title{MFP.
Our dummy package has a collection of functions for creating a greeting.}
\description{
MFP.
Our dummy package has a collection of functions for creating a greeting.
}
